// api/request/user.request.ts
export class UserRequest{

  firstName: string;
	lastName: string;
	userName: string;
	securityRole: string;
	userCategory: string;
	securityRoleDisplayName: string;
	userType: string;
	bpList: string[];
	billingAce: string;
	startStopAce: string;
	licenseNum: string;
	email: string;
	phoneNum: string;
	phoneNumExtn: string;
	altPhoneNum: string;
	altContactPhExtn: string;
    faxNum: string;
	businessName: string;
	address: Address;
	ableToPayBills: boolean;
	ableToStartStopService: boolean;
	webSecurityRole: string;
	brokerBPNumber: string;
	sapId: string;
  password:string;

    bpNumber : string;
    bcadminUserName : string;
    brokerAdminLastname: string;
    brokerBpNumber: string;
    searchCriteria: string;
    customerAdminLastname:string;
    bpNumberList:string[];
    bpNameList:string[];
    bpAddressList:string[];
    hierarchyLevelList:string[];
    startStopAceList:string[];
    billingAceList:string[];
    agentType: string;
    fieldIndicator:string;
    fieldValue1:string;
    fieldValue2:string;
    portal:string;
    strUserName:string;
    strCallType:string;
    searchString:string;
    adminUserName:string;
    strTransactionType:string;
    strVUEEmailOne:string;
    strVUEEmailTwo:string;
    strVUEESIID:string;
    strVUEPropertyName:string;
    strVUEUnit:string;
    strVUEUnitType:string;
    strVUEUnitThreshold:string;
    strVUEStreetAddress:string;
    strVUECity:string;
    strVUEState:string;
    strVUEZipcode:string;
    strActivityPerformed:string;
    strActivityType:string;
}

class Address{
 streetName : string;
 streetNum: string;
 state: string;
 city: string;
 poBox: string;
 zipcode: string;

}
// ...
export class SearchUsersForUpdateRequest extends BaseRequest {

      searchString: string;
      searchCriteria: string;
      webSecurityRole: string;
      userCategory: string;
      userType: string;
}
// ...
export class ResetPwdDtlsRequest extends BaseRequest{
  txnId: string;
  userName: string;
  expirationDate:string;

}
// ...
export class UserProfileWithBPHierarchyRequest extends BaseRequest {
    strUserName: string;
}


export class UpdateUserRequest extends BaseRequest {
    userName: string;
    fieldIndicator: string;
    fieldValue1: string;
    fieldValue2: string;
    portal: string;
    bpNumber: string;
}


export class DeactivateUserRequest extends BaseRequest{

      strUserName:string;

  }

export class DissociateCustomerFromUserRequest extends BaseRequest {
    userName: string;
    bpNumberList: string[];
}